app/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
websockets>=12
python-multipart>=0.0.9
httpx>=0.27
pyyaml>=6
python-dotenv>=1.0
bcrypt>=4.0
claude-agent-sdk>=0.1
tenacity>=8.2
PyGithub>=2.0
anthropic>=0.40
